+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- UI: Adjust the color of names in message bubbles to have better contrast
+
 ### Security
 
 ## 4.0.0